% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoffsr.R
\name{knockoff_setup}
\alias{knockoff_setup}
\title{Setup knockoffsr}
\usage{
knockoff_setup(julia_dir, pkg_check = TRUE, ...)
}
\arguments{
\item{julia_dir}{string, path to folder containing Julia executable}

\item{pkg_check}{logical, check if Knockoffs.jl package exist and install if necessary}

\item{...}{Parameters are passed down to JuliaCall::julia_setup}
}
\description{
This function initializes Julia and the Knockoffs.jl package.
This will install Julia and the required packages
if they are missing.
}
\examples{

knockoffsr::knockoff_setup()

}
